\docType{methods}
\name{c,wishart-method}
\alias{c,inverseWishart,inverseWishart-method}
\alias{c,inverseWishart-method}
\alias{c,wishart,wishart-method}
\alias{c,wishart-method}
\title{Combine Values into a vector or list}
\usage{
\S4method{c}{wishart}(x, ...)

\S4method{c}{inverseWishart}(x, ...)
}
\arguments{
  \item{...}{objects to be concatenated.}
}
\description{
Concatenate Wishart objects

Concatenate inverse Wishart objects
}

