\name{euclideanMatrixDistance}
\alias{euclideanMatrixDistance}
\title{euclideanMatrixDistance}
\usage{
euclideanMatrixDistance(m1, m2)
}
\arguments{
  \item{m1}{the first matrix}

  \item{m2}{the second matrix}
}
\value{
the Euclidean distance between the matrices
}
\description{
Calculate the euclidean distance between two matrices of
equal dimension
}

