\docType{class}
\name{inverseWishart}
\alias{inverseWishart}
\title{wishart}
\description{
Class representing the inverse Wishart distribution
}
\section{Slots}{

\describe{
\item{\code{df}}{The degrees of freedom for the Wishart distribution.  Must be a positive integer.}

\item{\code{precision}}{The precision matrix of the inverse Wishart distribution.  Must be a square, positive
definite matrix.}
}}
\note{
For theoretical details, please see

Gupta, A. K., & Nagar, D. K. (2000). Matrix variate
distributions. Boca Raton, FL: Chapman & Hall.
}

