\name{isPositiveDefinite}
\alias{isPositiveDefinite}
\title{isPositiveDefinite}
\usage{
isPositiveDefinite(x, tol = 1e-18)
}
\arguments{
  \item{x}{a \code{matrix} object}

  \item{tol}{a value indicating the smallest integer
  considered positive}
}
\value{
TRUE if the matrix is positive definite, and FALSE
otherwise
}
\description{
check if a matrix is positive definite
}
\keyword{internal}

