\name{rInverseWishart}
\alias{rInverseWishart}
\title{rInverseWishart}
\usage{
rInverseWishart(invWishart, n = 1000)
}
\arguments{
  \item{invWishart}{the inverse Wishart distribution}

  \item{n}{the number of samples}
}
\value{
a sample of matrices from the inverse Wishart distribution
}
\description{
Generate a sample from an inverse Wishart distribution.
}
\seealso{
\code{\link{inverseWishart}}
}

