\name{simulate.sumInvWisharts}
\alias{simulate.sumInvWisharts}
\title{simulate.sumInvWisharts}
\usage{
\method{simulate}{sumInvWisharts}(invWishartList, replicates = 1000)
}
\arguments{
  \item{invWishartList}{list of inverse Wisharts}

  \item{replicates}{the number of samples to generate}
}
\value{
A sample of matrices from the distribution of the sum of
the inverse Wishart matrices
}
\description{
Generate a sample of matrices from the distribution of the
sum of inverse Wishart matrices
}

