\name{traceVariance}
\alias{traceVariance}
\title{traceVariance}
\usage{
traceVariance(invWishart)
}
\arguments{
  \item{invWishart}{the inverse Wishart object}
}
\value{
the variance of the trace of the specified inverse Wishart
}
\description{
Calculate the variance of the trace of an inverse Wishart
matrix
}
\seealso{
\code{\link{inverseWishart}}
}

